#235 [fix] Option에서 LocalDataSource와 RemoteDataSource 분리
