+++ app/src/main/java/com/gojol/notto/ui/option/OptionViewModel.kt
@@ -1,6 +1,5 @@
 package com.gojol.notto.ui.option
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
